refactor: extract core domain module
