+++ apps/core/domain/entity.go
