+++ apps/core/domain/registry.go
