+++ apps/core/go.mod
@@ -1,3 +1,10 @@
 module github.com/azusachino/felicia/apps/core
 
 go 1.26
+
+require (
+	github.com/google/go-cmp v0.7.0
+	github.com/google/uuid v1.6.0
+	github.com/paulmach/orb v0.13.0
+	gopkg.in/yaml.v3 v3.0.1
+)
