+++ cmd/api/main.go
@@ -11,10 +11,10 @@ import (
 	"github.com/jackc/pgx/v5/pgxpool"
 
 	"github.com/azusachino/felicia"
+	"github.com/azusachino/felicia/apps/core/domain"
 	"github.com/azusachino/felicia/internal/api"
 	"github.com/azusachino/felicia/internal/config"
 	"github.com/azusachino/felicia/internal/dawarich"
-	"github.com/azusachino/felicia/internal/domain"
 	"github.com/azusachino/felicia/internal/immich"
 	"github.com/azusachino/felicia/internal/importer"
 	"github.com/azusachino/felicia/internal/store/pg"
