+++ internal/api/server.go
@@ -16,7 +16,7 @@ import (
 	"github.com/google/uuid"
 	"github.com/paulmach/orb"
 
-	"github.com/azusachino/felicia/internal/domain"
+	"github.com/azusachino/felicia/apps/core/domain"
 	"github.com/azusachino/felicia/internal/importer"
 	"github.com/azusachino/felicia/internal/publication"
 )
