+++ internal/api/server_test.go
@@ -15,8 +15,8 @@ import (
 	"github.com/google/uuid"
 	"github.com/paulmach/orb"
 
+	"github.com/azusachino/felicia/apps/core/domain"
 	"github.com/azusachino/felicia/internal/api"
-	"github.com/azusachino/felicia/internal/domain"
 	"github.com/azusachino/felicia/internal/importer"
 )
