+++ internal/publication/projection.go
@@ -6,7 +6,7 @@ import (
 	"github.com/google/uuid"
 	"github.com/paulmach/orb"
 
-	"github.com/azusachino/felicia/internal/domain"
+	"github.com/azusachino/felicia/apps/core/domain"
 )
 
 // JourneyListItem is the compact landing-page projection for one journey.
